requirements.txt
# ==============================================================================
# STELLA AGENT - Dependências do Projeto
# ==============================================================================

# Reconhecimento facial ---------------------------------------
deepface==0.0.92               # Reconhecimento facial (compatível Windows)
opencv-python>=4.5.0,<5.0.0    # Processamento de imagem
tf_keras==2.19.0               # Interface Keras para TensorFlow
# -------------------------------------------------------------

# Configuração e ambiente -------------------------------------
python-dotenv==1.1.1           # Variáveis de ambiente (.env)
pyyaml==6.0.2                  # Arquivos de configuração YAML
# -------------------------------------------------------------

# Logging -----------------------------------------------------
loguru==0.7.3                  # Sistema de logging avançado
# -------------------------------------------------------------

# Desenvolvimento ---------------------------------------------
pytest==8.4.1                  # Framework de testes
fastapi                        # Framework web para Python
uvicorn                       # Servidor ASGI para FastAPI
# -------------------------------------------------------------

# Comunicações (AMQP/WebSocket) -----------------------------------
pika==1.3.2                    # Cliente oficial RabbitMQ para Python
pusher==3.3.3                  # Cliente oficial Pusher para Python
# -------------------------------------------------------------

# Processamento de Linguagem Natural (NLP) --------------------
google-generativeai>=0.8.0     # API Gemini
spacy>=3.7.0                   # Suporte híbrido para NLP em português
# -------------------------------------------------------------

# Dependências fixas para compatibilidade ---------------------
numpy>=1.25.2,<2.0.0          # Compatível com DeepFace e Gemini
protobuf>=4.25.3,<5.0.0       # Compatível com ambos
tensorflow>=2.15.0            # Backend para DeepFace
transformers                   # Para zero-shot e NER (versão mais recente)
torch                          # Backend para transformers (versão mais recente)
requests                       # Para fazer requisições HTTP
python-multipart
# -------------------------------------------------------------

# ==============================================================================
# Para instalar: pip install -r requirements.txt após seguir INSTALLATION_GUIDE
# ==============================================================================